Added context for Todo Items
In the ApplicationDbContext I added the table for the todo items, then i migrated the changes to the database. I configured the service for the actual items that will be stored in the database and added a form in the index that allows you to add new todo items to the database.